Fixed problem with standing queue implementation
Fixed problem with standing queue dequeueing too
many elements.  Brain node now publishes the
position all the time to prevent any dropped
packets.  Control node publishes that it wants
to get a new position all the time for the same
reason.  Control node makes sure that if the
same goal position is sent, it ignores it and
continues moving to the goal.  Brain node, if
it sees that control has finished and wants a
new position, dequeues a new position and
publishes and then ignores requests for a new
positions for one second, in case control node
sends multiple position requests.